requirements.txt
# Frontend
streamlit==1.50.0

# Backend APIs
pillow==11.3.0
pytesseract==0.3.13          # local OCR fallback
requests==2.32.5             # for Azure Vision API calls

# Gemini
google-generativeai==0.8.3   # official Gemini SDK

# Notion
notion-client==2.2.1

# Env management
python-dotenv==1.1.1

# MCP server
fastapi==0.115.5
uvicorn==0.32.0
mcp>=0.1.4
